Add the pkcs11-sys crate
The pkcs11-sys crate directly uses the new bindgen feature available in
0.56.0 to generate bindings and minimal library code to dynamically load
the PKCS11 library.
The crate commits in tree the bindings for the x86_64-unknown-linux-gnu
target with the possibility of generating them during build time through
a feature. The goal is to make compilation easier and faster for
supported targets.
This crate exposes all values and types coming from the interface.

Co-authored-by: Joe Ellis <joe.ellis@arm.com>
Co-authored-by: Ionut Mihalcea <ionut.mihalcea@arm.com>
Co-authored-by: Hugues de Valon <hugues.devalon@arm.com>
Signed-off-by: Hugues de Valon <hugues.devalon@arm.com>