Add a new safe and idiomatic Rust abstraction
This commit adds the `new` module in the pkcs11 library. This new module
is an abstraction directly over the pkcs11-sys crate.
It currently only abstracts a subset of the PKCS11 interface but all the
types and functions abstracted are idiomatic and safe. The functions
only work with abstracted types and do not use raw bindgen types.
This commit also adds a few tests for this new module and also
conversions from PSA Crypto Algorithm type.

Co-authored-by: Joe Ellis <joe.ellis@arm.com>
Co-authored-by: Ionut Mihalcea <ionut.mihalcea@arm.com>
Co-authored-by: Hugues de Valon <hugues.devalon@arm.com>
Signed-off-by: Hugues de Valon <hugues.devalon@arm.com>